requirements.txt
edgartools
yfinance
fastapi
uvicorn
spacy
beautifulsoup4
lxml
cssutils
pytest
